backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.7
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==3.2.16
django-colorfield==0.11.0
django-cors-headers==3.13.0
django-filter==23.5
django-import-export==3.3.6
django-templated-mail==1.1.1
djangorestframework==3.12.4
djoser==2.1.0
et-xmlfile==1.1.0
flake8==6.0.0
flake8-isort==6.0.0
gunicorn==20.1.0
idna==3.6
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.3
MarkupPy==1.14
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.2.0
psycopg2-binary==2.9.3
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.5.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0
